functions/graph/shared/requirements.txt
# Shared dependencies
google-cloud-firestore==2.14.0
google-cloud-secret-manager==2.18.0
google-cloud-tasks==2.15.0
neo4j==5.15.0
google-generativeai==0.3.2
